test State
A common way to inform React of a data change is by calling
setState(data, callback). This method merges data into this.state and
re-renders the component. When the component finishes re-rendering, the
optional callback is called. Most of the time you'll never need to
provide a callback since React will take care of keeping your UI
up-to-date for you.